feat - fix page wallet
